requirements.txt
PyYAML>=6.0
pyinstaller>=6.0

# Testing
pytest>=7.0
pytest-cov>=4.0
